Module narwhal_primary::consensus

Structs

Bullshark

ChannelMetrics

Consensus

ConsensusMetrics

ConsensusRound

Holds information about a committed round in consensus. When a certificate gets committed then the corresponding certificate’s round is considered a “committed” round. It bears both the committed round and the corresponding garbage collection round.

ConsensusState

The state that needs to be persisted for crash-recovery.

LeaderSchedule

The LeaderSchedule is responsible for producing the leader schedule across an epoch. It provides methods to derive the leader of a round based on the provided leader swap table. This struct can be cloned and shared freely as the internal parts are atomically updated.

LeaderSwapTable

Enums

ConsensusError

Outcome

Constants

DEFAULT_CHANNEL_SIZE

The default channel size used in the consensus and subscriber logic.

NUM_SHUTDOWN_RECEIVERS

The number of shutdown receivers to create on startup. We need one per component loop.

Functions

gc_round

Calculates the GC round given a commit round and the gc_depth

make_consensus_store

Type Aliases

Dag

The representation of the DAG in memory.