Crate narwhal_primary

Modules

consensus

Structs

Primary

PrimaryChannelMetrics

Constants

CHANNEL_CAPACITY

The default channel capacity for each channel of the primary.

NUM_SHUTDOWN_RECEIVERS

The number of shutdown receivers to create on startup. We need one per component loop.