Crate narwhal_primary

Re-exports

pub use crate::block_synchronizer::mock::MockBlockSynchronizer;

pub use crate::block_synchronizer::BlockHeader;

Modules

block_synchronizer

Structs

BlockRemover

BlockRemover is responsible for removing blocks identified by their certificate digest from across our system. On high level It will make sure that the DAG is updated, internal storage where there certificates and headers are stored, and the corresponding batches as well.

BlockWaiter

BlockWaiter is responsible for fetching the block data from the downstream worker nodes. A block is basically the aggregate of batches of transactions for a given certificate.

EndpointMetrics

GetBlockResponse

Primary

PrimaryChannelMetrics

Constants

CHANNEL_CAPACITY

The default channel capacity for each channel of the primary.

NUM_SHUTDOWN_RECEIVERS

The number of shutdown receivers to create on startup. We need one per component loop.